initial refactor for legacy and start pacakges
