--- lib/nuxt-start.js
@@ -1,4 +1,0 @@
-import core from './core'
-import * as Utils from './common/utils'
-
-export default Object.assign({ Utils }, core)
