+++ packages/nuxt-legacy/README.md
@@ -1,0 +1,3 @@
+# nuxt-legacy
+
+> Legacy build of Nuxt.js for Node.js < 8.0.0
