+++ packages/nuxt-start/src/index.js
@@ -1,0 +1,4 @@
+import core from '../../../lib/core'
+import * as Utils from '../../../lib/common/utils'
+
+export default Object.assign({ Utils }, core)
