--- scripts/rollup/nuxt-start.js
@@ -1,6 +1,0 @@
-import config from './rollup.config'
-
-export default config({
-  name: 'nuxt-start',
-  input: './lib/nuxt-start.js'
-})
