--- scripts/rollup/nuxt.js
@@ -1,6 +1,0 @@
-import config from './rollup.config'
-
-export default config({
-  name: 'nuxt',
-  input: './lib/nuxt.js'
-})
