--- start/index.js
@@ -1,8 +1,0 @@
-/*!
- * Nuxt.js
- * (c) 2016-2018 Chopin Brothers
- * Core maintainers: Pooya Parsa (@pi0) - Clark Du (@clarkdo)
- * Released under the MIT License.
- */
-
-module.exports = require('./dist/nuxt-start.js')
