--- start/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "nuxt-start",
-  "version": "2.0.0",
-  "description": "runtime-only build for nuxt",
-  "contributors": [
-    {
-      "name": "Sebastien Chopin (@Atinux)"
-    },
-    {
-      "name": "Alexandre Chopin (@alexchopin)"
-    },
-    {
-      "name": "Pooya Parsa (@pi0)"
-    },
-    {
-      "name": "Clark Du (@clarkdo)"
-    }
-  ],
-  "main": "index.js",
-  "module": "./lib/nuxt.js",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nuxt/nuxt.js"
-  },
-  "files": [
-    "bin",
-    "lib",
-    "dist"
-  ],
-  "keywords": [
-    "nuxt",
-    "nuxt.js",
-    "nuxtjs",
-    "vue",
-    "vue.js",
-    "vuejs",
-    "vue universal",
-    "vue ssr",
-    "vue isomorphic",
-    "vue versatile"
-  ],
-  "homepage": "https://github.com/nuxt/nuxt.js#readme",
-  "bin": {
-    "nuxt-start": "./bin/nuxt-start"
-  },
-  "engines": {
-    "node": ">=8.0.0",
-    "npm": ">=5.0.0"
-  },
-  "dependencies": {
-  },
-  "collective": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/nuxtjs",
-    "logo": "https://opencollective.com/nuxtjs/logo.txt?reverse=true&variant=variant2"
-  }
-}
